Rename tool-utils.ts to utils.ts
