+++ extensions/serpapi-search/src/utils.ts
